fixed cohort.subject_id data type

+++ source/org.ohdsi.cdm.framework.etl/org.ohdsi.cdm.framework.etl.common/ETL/Common/Scripts/MSSQL/v6.0/CreateTables.sql
@@ -488,7 +488,7 @@ CREATE TABLE {sc}.cdm_source
 
 create table {sc}.cohort (
 			cohort_definition_id bigint not null,
-			subject_id integer not null,
+			subject_id bigint not null,
 			cohort_start_date date not null,
 			cohort_end_date date not null );
