+++ source/org.ohdsi.cdm.framework.etl/org.ohdsi.cdm.framework.etl.common/ETL/Common/Scripts/MySql/v6.0/CreateTables.sql
@@ -488,7 +488,7 @@ CREATE TABLE {sc}.cdm_source
 
 CREATE TABLE {sc}.cohort (
 			cohort_definition_id bigint(20) NOT NULL,
-			subject_id int(11) NOT NULL,
+			subject_id bigint(20) NOT NULL,
 			cohort_start_date date NOT NULL,
 			cohort_end_date date NOT NULL );
